ossfuzz: silence picky compiler warnings
```
tests/ossfuzz/ssh2_client_fuzzer.cc:41:23: error: empty expression statement has no effect; remove unnecessary ';' to silence this warning [-Werror,-Wextra-semi-stmt]
  FUZZ_ASSERT(rc == 0);
                      ^
tests/ossfuzz/ssh2_client_fuzzer.cc:45:14: error: comparison of integers of different signs: 'ssize_t' (aka 'long') and 'size_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
  if(written != size)
     ~~~~~~~ ^  ~~~~
```

```
tests/ossfuzz/standaloneengine.cc:35:20: warning: implicit conversion changes signedness: 'long' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
      buffer_len = ftell(infile);
                 ~ ^~~~~~~~~~~~~
```

+++ tests/ossfuzz/standaloneengine.cc
@@ -32,7 +32,7 @@ int main(int argc, char **argv)
 
       /* Get the length of the file. */
       fseek(infile, 0L, SEEK_END);
-      buffer_len = ftell(infile);
+      buffer_len = (size_t)ftell(infile);
 
       /* Reset the file indicator to the beginning of the file. */
       fseek(infile, 0L, SEEK_SET);
